fix: address Copilot review comments on PR #45
- Hide skip button when storeUrl is null for update-type announcements
- Extract shared GradientButton widget to avoid duplication
- Fix version-check CI to explicitly fetch base ref before comparison
