+++ frontend/lib/global/widgets/gradient_button.dart
@@ -1,0 +1,45 @@
+import 'package:flutter/material.dart';
+
+class GradientButton extends StatelessWidget {
+  const GradientButton({
+    super.key,
+    required this.gradient,
+    required this.label,
+    required this.onTap,
+  });
+
+  final LinearGradient gradient;
+  final String label;
+  final VoidCallback onTap;
+
+  @override
+  Widget build(BuildContext context) {
+    return SizedBox(
+      width: double.infinity,
+      height: 52,
+      child: DecoratedBox(
+        decoration: BoxDecoration(
+          gradient: gradient,
+          borderRadius: BorderRadius.circular(14),
+        ),
+        child: Material(
+          color: Colors.transparent,
+          child: InkWell(
+            borderRadius: BorderRadius.circular(14),
+            onTap: onTap,
+            child: Center(
+              child: Text(
+                label,
+                style: const TextStyle(
+                  color: Colors.white,
+                  fontWeight: FontWeight.bold,
+                  fontSize: 16,
+                ),
+              ),
+            ),
+          ),
+        ),
+      ),
+    );
+  }
+}
